:truck: chore(api) move read model to real context
