+++ api/tests/integration/infrastructure/repositories/certification-point-of-contact.repository.test.js
@@ -1,9 +1,9 @@
 import * as centerRepository from '../../../../src/certification/enrolment/infrastructure/repositories/center-repository.js';
+import { AllowedCertificationCenterAccess } from '../../../../src/identity-access-management/domain/read-models/AllowedCertificationCenterAccess.js';
 import * as certificationPointOfContactRepository from '../../../../src/identity-access-management/infrastructure/repositories/certification-point-of-contact.repository.js';
 import { Organization } from '../../../../src/organizational-entities/domain/models/Organization.js';
 import { NotFoundError } from '../../../../src/shared/domain/errors.js';
 import { CertificationCenter } from '../../../../src/shared/domain/models/CertificationCenter.js';
-import { AllowedCertificationCenterAccess } from '../../../../src/shared/domain/read-models/AllowedCertificationCenterAccess.js';
 import { catchErr, databaseBuilder, domainBuilder, expect } from '../../../test-helper.js';
 
 describe('Integration | Identity Access Management |  Repository | CertificationPointOfContact', function () {
